:chart_with_upwards_trend: Injecte la dépendance à Piwik dans entry
