+++ source/entry.dev.js
@@ -2,7 +2,9 @@ import React from 'react'
 import { render } from 'react-dom'
 import { compose, createStore, applyMiddleware } from 'redux'
 import reducers from './reducers/reducers'
-import DevTools from './DevTools'
+import DevTools from '../DevTools'
+import { Provider } from 'react-redux'
+import Layout from './Layout'
 import { AppContainer } from 'react-hot-loader'
 import debounceFormChangeActions from './debounceFormChangeActions'
 import computeThemeColours from './components/themeColours'
@@ -10,8 +12,6 @@ import { getIframeOption, getUrl } from './utils'
 
 import { rules, rulesFr } from 'Engine/rules'
 import lang from './i18n'
-
-import App from './containers/App.dev'
 
 let initialStore = {
 	iframe: getUrl().includes('iframe'),
@@ -23,15 +23,29 @@ let enhancer = compose(
 	DevTools.instrument({ maxAge: 10 })
 )
 
+let tracker = {
+	push: console.log,
+	connectToHistory: () => {}
+}
+
 let initialRules = lang == 'en' ? rules : rulesFr
-let store = createStore(reducers(initialRules), initialStore, enhancer)
+let store = createStore(reducers(tracker, initialRules), initialStore, enhancer)
 let anchor = document.querySelector('#js')
+
+let App = ({ store }) => (
+	<Provider store={store}>
+		<div id="dev">
+			<Layout />
+			<DevTools />
+		</div>
+	</Provider>
+)
 
 render(<App store={store} />, anchor)
 
 // Hot react component reloading. Unstable but helpful.
 if (module.hot) {
-	module.hot.accept('./containers/App.dev', () => {
+	module.hot.accept('./containers/Layout', () => {
 		render(
 			<AppContainer>
 				<App store={store} />
